requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
langchain==0.3.7
langchain-community==0.3.5
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-chroma==0.1.4
chromadb==0.5.20
sentence-transformers==3.3.1
huggingface-hub==0.26.2
groq==0.11.0
pypdf==5.1.0
pydantic==2.9.2
